Found the problem
the noInterrupts() after sleep_disable is what breaks the sleeping mode.
Once commenting out all the sleep enables and saw it was working with delays led me to looks whats different between the sleep code with the working practice code and saw I added that noInnterrupts cause I thought it would be a good protection catch, but I guess not.